split sdk_a files
