+++ examples/get_next_greeting_msg.py
@@ -1,7 +1,7 @@
-import queue
 import sys
 import colink as CL
-from colink.sdk_a import CoLink, byte_to_str
+from colink import CoLink
+from colink.sdk_a import byte_to_str
 
 if __name__ == "__main__":
     addr = sys.argv[1]
