+++ examples/host_import_user.py
@@ -1,7 +1,7 @@
 import sys
 import logging
+from colink import CoLink
 from colink.sdk_a import (
-    CoLink,
     get_time_stamp,
     generate_user,
     prepare_import_user_signature,
