+++ examples/host_import_users_and_set_registry.py
@@ -1,9 +1,8 @@
 import sys
 import logging
-import os
 import colink as CL
+from colink import CoLink
 from colink.sdk_a import (
-    CoLink,
     get_time_stamp,
     generate_user,
     prepare_import_user_signature,
