+++ examples/user_start_protocol_operator.py
@@ -1,5 +1,6 @@
 import sys
-from colink.sdk_a import CoLink, decode_jwt_without_validation
+from colink import CoLink
+from colink.sdk_a import decode_jwt_without_validation
 
 if __name__ == "__main__":
     addr = sys.argv[1]
